requirements.txt
aiofiles==24.1.0
aiogram==3.0.0
python-dotenvaiohttp==3.11.18
anyio==4.9.0
blinker==1.9.0
fastapi==0.115.12
Flask==3.1.0
httpx==0.28.1
Jinja2==3.1.6
MarkupSafe==3.0.2
orjson==3.10.18
pydantic==2.11.4
python-telegram-bot==22.0
requests==2.32.3
starlette==0.46.2
uvicorn==0.34.2
